% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuousMastery.R
\name{continuousDCM}
\alias{continuousDCM}
\title{DCM Continuous Mastery}
\usage{
continuousDCM(I)
}
\arguments{
\item{I}{Number of examinees}
}
\description{
Creates response data for an experimental continuous mastery model
}
\section{\strong{Notation}}{

 \describe{
   \tabular{ll}{
     q \tab Q-matrix \cr
     I \tab Number of examinees \cr
     J \tab Number of items \cr
     K \tab Number of skills \cr
     alpha \tab Examinee skill mastery
   }
 }
}
\author{
Dave Rackham \email{ddrackham@gmail.com}
}
\references{
\url{http://onlinelibrary.wiley.com/doi/10.1002/j.2333-8504.2008.tb02157.x/abstract} # TODO: Update
}
\keyword{q-matrix}

