% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hartzRoussosData.R
\name{hartzRoussosData}
\alias{hartzRoussosData}
\title{Hartz Roussos (2008) Data simulation}
\usage{
hartzRoussosData()
}
\arguments{
\item{j}{Examinees}

\item{i}{Items}

\item{k}{skills}

\item{x}{response matrix}

\item{pi}{Probability that an examinee having mastered all the Q required skills for item i will correctly apply all the skills when solving item i}
}
\description{
Simulates the Hartz Roussos data from "The Fusion Model for Skills Diagnosis: Blending Theory with Practicality" (2008)
}
\examples{

}
\keyword{hartz}
\keyword{roussos}

