% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hartzRoussosQ.R
\name{hartzRoussosQ}
\alias{hartzRoussosQ}
\title{Hartz Roussos (2008) Q-matrix}
\usage{
hartzRoussosQ()
}
\description{
Simulates the Hartz Roussos Q-matrix from "The Fusion Model for Skills Diagnosis: Blending Theory with Practicality" (2008)
}
\examples{
q <- hartzRoussosQ()

}
\keyword{hartz}
\keyword{q-matrix}
\keyword{roussos}

