% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hartzRoussosQ.R
\name{hartzRoussosQLow}
\alias{hartzRoussosQLow}
\title{Hartz Roussos (2008) Q-matrix}
\usage{
hartzRoussosQLow()
}
\description{
Simulates the Hartz Roussos Q-matrix from "The Fusion Model for Skills Diagnosis: Blending Theory with Practicality" (2008)
}
\examples{
q <- hartzRoussosQLow()

}
\author{
Dave Rackham \email{ddrackham@gmail.com}
}
\references{
\url{http://onlinelibrary.wiley.com/doi/10.1002/j.2333-8504.2008.tb02157.x/abstract}
}
\keyword{hartz}
\keyword{q-matrix}
\keyword{roussos}

