% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1PLData.R
\name{IRT_1PL.DataSim}
\alias{IRT_1PL.DataSim}
\title{IRT 1PL Data Generator}
\usage{
IRT_1PL.DataSim(I, J)
}
\arguments{
\item{I}{Number of examinees}

\item{J}{Number of items}
}
\description{
Creates response data for a 1PL IRT Model
}
\section{\strong{Notation}}{

 \describe{
   \tabular{ll}{
     theta \tab Respondent ability parameter \cr
     beta \tab Item difficulty parameter \cr
     I \tab Number of respondents \cr
     J \tab Number of items \cr
   }
 }
}
\author{
Dave Rackham \email{ddrackham@gmail.com}
}
\keyword{1PL}
\keyword{IRT}

