% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDinaData.R
\name{rDINA}
\alias{rDINA}
\title{RDINA data simulation}
\usage{
rDINA()
}
\arguments{
\item{J}{Number of examinees}

\item{jj}{Student for observation n}

\item{I}{Number of items}

\item{ii}{Item for observation n}

\item{K}{Number of skills}
}
\description{
Creates a response data for the RDINA model
}
\examples{
data <- rDINA()

}
\author{
Dave Rackham \email{ddrackham@gmail.com}
}
\references{
\url{http://onlinelibrary.wiley.com/doi/10.1002/j.2333-8504.2008.tb02157.x/abstract}
}
\keyword{q-matrix}

