% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDinaDataHartzRoussosQ.R
\name{rDINAHartzRoussosQ}
\alias{rDINAHartzRoussosQ}
\title{RDINA data simulation with Hartz Roussos Q-matrix (low)}
\usage{
rDINAHartzRoussosQ(I)
}
\arguments{
\item{I}{Number of examinees}
}
\description{
Creates response data for the RDINA model and the Hartz Roussos Q-matrix (low)
}
\details{
#' @section \strong{Important notes}:
\describe{
 As of 5-20-2016 I am unsure if the data simulation algorithm is accurate or effective. It needs to be connected to the simpleQ algorithm.
}
}
\section{\strong{Notation}}{

 \describe{
   \tabular{ll}{
     q \tab Q-matrix \cr
     I \tab Number of examinees \cr
     J \tab Number of items \cr
     K \tab Number of skills \cr
     f \tab False alarm rate \cr
     d \tab Item descrimination (detection) \cr
   }
 }
}
\author{
Dave Rackham \email{ddrackham@gmail.com}
}
\references{
\url{http://onlinelibrary.wiley.com/doi/10.1002/j.2333-8504.2008.tb02157.x/abstract} # TODO: Update
}
\keyword{q-matrix}

