% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDinaSimpleQ.R
\name{rDINASimpleQ}
\alias{rDINASimpleQ}
\title{RDINA data simulation}
\usage{
rDINASimpleQ(I)
}
\arguments{
\item{I}{Number of examinees}
}
\description{
Creates response data for the RDINA model
}
\section{\strong{Notation}}{

 \describe{
   \tabular{ll}{
     q \tab Q-matrix \cr
     I \tab Number of examinees \cr
     J \tab Number of items \cr
     K \tab Number of skills \cr
     f \tab False alarm rate \cr
     d \tab Item descrimination (detection) \cr
   }
 }
}
\author{
Dave Rackham \email{ddrackham@gmail.com}
}
\references{
\url{http://onlinelibrary.wiley.com/doi/10.1002/j.2333-8504.2008.tb02157.x/abstract} # TODO: Update
}
\keyword{q-matrix}

