% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUMSimpleQData.R
\name{RUMSimpleQData}
\alias{RUMSimpleQData}
\title{RUM Simple Q matrix data simulation}
\usage{
RUMSimpleQData()
}
\description{
Creates response data for a simplified version of the RUM model using the Simple 2 Attribute Q matrix
}
\details{
#' @section \strong{Important notes}:
\describe{
 As of 5-20-2016 I am unsure if the data simulation algorithm is accurate or effective. I abandoned this model in favor for the R-DINA.
}
}
\section{\strong{Notation}}{

 \describe{
   \tabular{ll}{
     JJ \tab Number of examinees \cr
     II \tab Number of items \cr
     KK \tab Number of skills \cr
     j \tab Examinee j \cr
     i \tab Item i \cr
     k \tab Skill k \cr
     alphaK \tab Skill mastery population proportion vector \cr
     alphaJK \tab Examinee skill mastery profile \cr
     x \tab response matrix \cr
     pi \tab Probability that an examinee having mastered all the Q required skills for item i will correctly apply all the skills when solving item i \cr
     iParamsLow \tab Matrix of item parameters for ideal low complexity model \cr
     rStar \tab Item discrimination \cr
     kappa \tab Mastery threshold parameter \cr
   }
 }
}
\examples{
data <- RUMSimpleQData()

}
\author{
Dave Rackham \email{ddrackham@gmail.com}
}
\references{
\url{http://onlinelibrary.wiley.com/doi/10.1002/j.2333-8504.2008.tb02157.x/abstract}
}
\keyword{q-matrix}

