% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleData.R
\name{simpleData}
\alias{simpleData}
\title{Simplified data simulation}
\usage{
simpleData()
}
\arguments{
\item{J}{Number of examinees}

\item{jj}{Student for observation n}

\item{I}{Number of items}

\item{ii}{Item for observation n}

\item{K}{Number of skills}

\item{alphaK}{Skill mastery population proportion vector}

\item{alphaJK}{Examinee skill mastery profile}

\item{masteryJK}{Continuous probability of mastery}

\item{x}{response matrix}

\item{iParamsLow}{Matrix of item parameters for ideal low complexity model}

\item{rStar}{Item discrimination r}

\item{kappa}{Mastery threshold parameter}
}
\description{
Creates response data for a simplified version of the RUM model using the Simple 2 Attribute Q matrix
}
\examples{
data <- simpleData()

}
\author{
Dave Rackham \email{ddrackham@gmail.com}
}
\references{
\url{http://onlinelibrary.wiley.com/doi/10.1002/j.2333-8504.2008.tb02157.x/abstract}
}
\keyword{q-matrix}

