% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleQ.R
\name{simpleQ}
\alias{simpleQ}
\title{Simple 2 Attribute Q-matrix}
\usage{
simpleQ()
}
\description{
Simple 2 Attribute Q-matrix
}
\examples{
q <- simpleQ()

}
\author{
Dave Rackham \email{ddrackham@gmail.com}
}
\references{
\url{http://onlinelibrary.wiley.com/doi/10.1002/j.2333-8504.2008.tb02157.x/abstract}
}
\keyword{q-matrix}

