% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hartzRoussosData.R
\name{hartzRoussosData}
\alias{hartzRoussosData}
\title{Hartz Roussos (2008) Data simulation}
\usage{
hartzRoussosData()
}
\arguments{
\item{JJ}{Number of examinees}

\item{j}{Examinee j}

\item{II}{Number of items}

\item{i}{Item i}

\item{KK}{Number of skills}

\item{k}{Skill k}

\item{alphaK}{Skill mastery proportion vector}

\item{alphaJK}{Examinee skill mastery profile}

\item{x}{response matrix}

\item{pi}{Probability that an examinee having mastered all the Q required skills for item i will correctly apply all the skills when solving item i.}

\item{iParamsLow}{Matrix of item parameters for ideal low complexity model}

\item{r}{Item discrimination r}

\item{c}{Item difficulty c}
}
\description{
Simulates the Hartz Roussos data from "The Fusion Model for Skills Diagnosis: Blending Theory with Practicality" (2008)
}
\examples{

}
\keyword{hartz}
\keyword{roussos}

